put myst_invoke() into invoke.c
